Add "ready" to bootmarker.
Otherwise, the module may reset if you send it another command before the reset completes (and the "ready" prints). I was seeing the ATE0 command (disable echo) getting sent before the module finished resetting, which seemed to trigger another reset. The module would then print another boot message, but not the "OK" that the library was expecting as the result of the ATE0 command.